+++ src/factory/taskName.ts
@@ -27,10 +27,6 @@ enum TaskName {
      */
     SettleCreditCard = 'settleCreditCard',
     /**
-     * 注文作成
-     */
-    CreateOrder = 'createOrder',
-    /**
      * 返品タスク
      */
     ReturnOrder = 'returnOrder',