+++ src/index.ts
@@ -19,7 +19,6 @@ import * as TaskFactory from './factory/task';
 import * as AggregateEventReservationsTaskFactory from './factory/task/aggregateEventReservations';
 import * as CancelCreditCardTaskFactory from './factory/task/cancelCreditCard';
 import * as CancelSeatReservationTaskFactory from './factory/task/cancelSeatReservation';
-import * as CreateOrderTaskFactory from './factory/task/createOrder';
 import * as CreatePlaceOrderReportTaskFactory from './factory/task/createPlaceOrderReport';
 import * as CreateReturnOrderReportTaskFactory from './factory/task/createReturnOrderReport';
 import * as ReturnOrderTaskFactory from './factory/task/returnOrder';
@@ -86,7 +85,6 @@ export namespace task {
     export import aggregateEventReservations = AggregateEventReservationsTaskFactory;
     export import cancelCreditCard = CancelCreditCardTaskFactory;
     export import cancelSeatReservation = CancelSeatReservationTaskFactory;
-    export import createOrder = CreateOrderTaskFactory;
     export import createPlaceOrderReport = CreatePlaceOrderReportTaskFactory;
     export import createReturnOrderReport = CreateReturnOrderReportTaskFactory;
     export import returnOrder = ReturnOrderTaskFactory;