+++ CHANGELOG.md
@@ -10,9 +10,13 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/).
 
 ### Changed
 
+- update @cinerino/factory
+
 ### Deprecated
 
 ### Removed
+
+- CreateOrderタスクを削除
 
 ### Fixed